分布式基础篇完成

+++ gulimall-product/src/main/java/site/wanjiahao/gulimall/product/entity/AttrEntity.java
@@ -53,7 +53,7 @@ public class AttrEntity implements Serializable {
 	 * 逻辑删除字段
 	 */
 	@TableLogic(value = "1", delval = "0")
-	private Long enable;
+	private Integer enable;
 	/**
 	 * 所属分类
 	 */
